+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## Removed
 
 -   Removed Alpine v2 support in favor of Alpine v3, use v1.0.0 for Alpine v2.
+-	Removed `subscribe()` and `updateSubscribers()` in favor of `Alpine.reactive()`.
 
 ## Changed
 
@@ -57,9 +58,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 -   First version
 
-[unreleased]: https://github.com/pinecone-router/router/compare/0.0.1...HEAD
+[unreleased]: https://github.com/pinecone-router/router/compare/2.0.0...HEAD
 [0.0.0]: https://github.com/pinecone-router/router/compare/0.0.0...0.0.0
 [0.0.1]: https://github.com/pinecone-router/router/compare/0.0.0...0.0.1
 [0.0.2]: https://github.com/pinecone-router/router/compare/0.0.1...0.0.2
 [0.0.3]: https://github.com/pinecone-router/router/compare/0.0.2...0.0.3
 [1.0.0]: https://github.com/pinecone-router/router/compare/0.0.3...1.0.0
+[2.0.0]: https://github.com/pinecone-router/router/compare/1.0.0...2.0.0